+++ projects/coffee/src/lib/models/index.ts
@@ -1,1 +1,2 @@
 export * from './coffee-model';
+export * from '../auth/models/coffee-auth-social';